feat: size legend for common situation, extract the sizeCircle legend
