+++ package.json
@@ -61,7 +61,7 @@
     "ci": "npm run lint && npm run test",
     "compress": "uglifyjs -c -m -o dist/component.min.js -- build/component.js",
     "coverage": "npm run coverage-generator && npm run coverage-viewer",
-    "coverage-generator": "torch --compile --coverage --renderer --source-pattern src/*.js,src/**/*.js --recursive test/unit/legend/",
+    "coverage-generator": "torch --compile --coverage --renderer --source-pattern src/*.js,src/**/*.js --recursive test/unit/legend/size-spec.js",
     "coverage-viewer": "torch-coverage",
     "demos": "electron ./demos/app.js",
     "demos-web": "node ./demos/app.js --web --port 2046",
@@ -77,7 +77,7 @@
     "test-all": "npm --compile run test && npm run test-bugs",
     "test-bugs": "torch --compile --renderer --recursive ./test/bugs",
     "test-bugs-live": "torch --compile --interactive --watch --recursive ./test/bugs",
-    "test-live": "torch --compile --interactive --watch --recursive ./test/unit/legend/",
+    "test-live": "torch --compile --interactive --watch --recursive ./test/unit/legend/size-spec.js",
     "watch": "webpack --config webpack-dev.config.js",
     "win-dev": "node ./bin/win-dev.js"
   },
